requirements.txt
py-cord
python-dotenv
wavelink
